feat : 비디오 컨트롤러 이모티콘 추가, 댓글 추가 개발

+++ src/main/java/com/kogo/kologbackend/application/emotion/dto/request/EmotionCreateRequest.java
@@ -3,6 +3,6 @@ package com.kogo.kologbackend.application.emotion.dto.request;
 public record EmotionCreateRequest(
         Long userId,
         Long logId,
-        int emotionId
+        Long emotionId
 ) {
 }
